update: main CLI app part 2
splitting functions and commands into UTILS
